requirements.txt
Django==2.2.4
djangorestframework==3.11.0
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
